tune_scvi now tracks marginal_ll
