+++ scvi/autotune/__init__.py
@@ -1,3 +1,12 @@
 from ._autotune import Autotune
+from ._callbacks import ModelSave, _TuneReportMetricFunctionsCallback
+from ._metrics import silhouette_metric
+from ._wrappers import tune_scvi
 
-__all__ = ["Autotune"]
+__all__ = [
+    "Autotune",
+    "tune_scvi",
+    "silhouette_metric",
+    "ModelSave",
+    "_TuneReportMetricFunctionsCallback",
+]
